pdf-saas/gemini-api-microservice/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai

